+++ CHANGELOG.md
@@ -28,6 +28,12 @@
   "INCLUDE_DELETED" option, migration task first posts the document, then deletes
   it in destination nodes.
 
+* Added new stats for ServiceHostManagementService to reflect the current size
+  of the authorization context Cache and the count of cache insertions.
+
+* Added support for request logging that logs all inbound requests to a Service
+  Host. Logging can be enabled or disabled during host start-up by calling
+  ServiceHost.setRequestLoggingInfo OR by making a PATCH to ServiceHostManagementService.
 
 ## 1.5.4